ref/requirements.txt
# FastAPI and web server
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.44
alembic==1.17.2

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1

# Pydantic
pydantic==2.12.4
pydantic-settings==2.12.0

# HTTP requests
requests==2.32.5
tenacity==9.1.2

# ML/NLP (optional but recommended)
scikit-learn==1.7.2
numpy==2.3.5

# Optional: Embeddings and Vector DB
sentence-transformers==5.1.2
qdrant-client==1.16.1
# torch==2.1.0

# Email validation
email-validator==2.3.0
